+++ src/routes/NoteViewer/NoteViewer.js
@@ -48,9 +48,29 @@ const NoteViewer = ({
 };
 
 NoteViewer.propTypes = {
-  history: PropTypes.object.isRequired,
-  location: PropTypes.object.isRequired,
-  match: PropTypes.object.isRequired,
+  history: PropTypes.shape({
+    goBack: PropTypes.func.isRequired,
+    replace: PropTypes.func.isRequired,
+  }).isRequired,
+  location: PropTypes.shape({
+    state: PropTypes.shape({
+      referredRecordData: PropTypes.shape({
+        instanceId: PropTypes.string.isRequired,
+        instanceTitle: PropTypes.string.isRequired,
+        itemBarcode: PropTypes.string.isRequired,
+        itemId: PropTypes.string.isRequired,
+        holdingsRecordId: PropTypes.string.isRequired,
+        requestCreateDate: PropTypes.string.isRequired,
+        requesterId: PropTypes.string.isRequired,
+        requesterName: PropTypes.string.isRequired,
+      }),
+    }),
+  }).isRequired,
+  match: PropTypes.shape({
+    params: PropTypes.shape({
+      noteId: PropTypes.string,
+    }),
+  }).isRequired,
 };
 
 export default NoteViewer;